KS-2-1 school refactor task .kt

+++ src/main/kotlin/lesson_2/KS-2-1 school.kt
@@ -1,6 +1,6 @@
 package org.example.lesson_2
 
 fun main() {
-    val schoolAssessment = (3 + 4 + 3 + 5) / 4.0
+    val schoolAssessment = (3 + 4 + 3 + 5).toFloat() / 4
     println(schoolAssessment)
 }
